requirements.txt
accelerate==1.12.0
datasets==4.4.1
einops==0.8.1
filelock==3.20.0
huggingface-hub==0.36.0
jaxtyping==0.3.3
lightning==2.6.0
numpy==2.3.5
psutil==7.1.3
PyQt6==6.10.0
PyYAML==6.0.3
seaborn==0.13.2
sentencepiece==0.2.1
torch==2.9.1 --index-url https://download.pytorch.org/whl/cu128
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.51.3
wandb==0.23.1
